Updates to the firedrake-rayleigh-benard solve. Not working yet.
I think this is the exact same configuration as the DOLFIN code.
But

(a) firedrake gives a different residual norm to dolfin;
(b) the first deflated solve converges to something slightly
different (but functionally the same branch), instead of being
forced away and finding the other solution like it should.