Make the scheduling strategy much more conservative. Hopefully with this we will not discover any branches twice. This comprises two changes:
(i) if we are thinking about sending out a
deflation task, and there are any *continuation* tasks that haven't
reached us yet, defer. This is because the continuation task will
almost surely have information for us (a solution to deflate). This
also ensures we won't discover this branch again.

(ii) if we have multiple deflation tasks in flight, and one of them
succeeds, invalidate all of the others: even if they report
success ignore them and send them out again. (If they report failure
then don't send them out again.) This ensures that we won't have
two deflations in flight that lock on to the same branch.